Former-commit-id: 81f57c76eb61220b762956856571b523700743c5 [formerly bdc83103f2e6e4497bb74bb6e634e8029aeb56cd]
Former-commit-id: 420fba5d0fbd0c296d3a530eccb2fee2442acfbd

+++ openquake/hazardlib/stats.py
@@ -97,7 +97,7 @@ def compute_pmap_stats(pmaps, stats, weights):
         assert pmap.shape_y == L, (pmap.shape_y, L)
     if len(sids) == 0:
         raise ValueError('All empty probability maps!')
-    sids = numpy.array(sorted(sids), numpy.float32)
+    sids = numpy.array(sorted(sids), numpy.uint32)
     nstats = len(stats)
     curves = numpy.zeros((len(pmaps), len(sids), L), numpy.float64)
     for i, pmap in enumerate(pmaps):
